requirements.txt
pymupdf==1.24.2
customtkinter==5.2.2
pillow==10.3.0
